Move convertToArray method to ResultParameterProcessor
